Moved evaluation stuff to seperate file.
